+++ addon/components/bs-table-pagination/table-footer.js
@@ -1,7 +1,7 @@
-import Component from '@ember/component';
-import layout from '../../templates/components/bs-table-pagination/table-footer';
+import Component from '@ember/component'
+import layout from '../../templates/components/bs-table-pagination/table-footer'
 
 export default Component.extend({
   layout,
   classNames: ['row', 'bs-table-pagination-footer']
-});
+})