+++ addon/components/light-table-pagination/table-actions.js
@@ -1,8 +1,7 @@
-import Component from '@ember/component';
-import layout from '../../templates/components/light-table-pagination/table-actions';
+import Component from '@ember/component'
+import layout from '../../templates/components/light-table-pagination/table-actions'
 
 export default Component.extend({
   layout,
   tagName: ''
-});
-
+})