+++ addon/components/light-table/columns/quick-filter-toggle.js
@@ -1,6 +1,6 @@
-import { isPresent } from '@ember/utils';
-import layout from '../../../templates/components/light-table/columns/quick-filter-toggle';
-import ColumnsBase from 'ember-light-table/components/columns/base';
+import { isPresent } from '@ember/utils'
+import layout from '../../../templates/components/light-table/columns/quick-filter-toggle'
+import ColumnsBase from 'ember-light-table/components/columns/base'
 
 export default ColumnsBase.extend({
   layout,
@@ -10,11 +10,11 @@ export default ColumnsBase.extend({
   actions: {
     toggleFilters () {
       this.get('table.visibleColumns').forEach((col) => {
-        let tpColumn = col.get('tpColumn');
+        let tpColumn = col.get('tpColumn')
         if (isPresent(tpColumn)) {
-          tpColumn.toggleProperty('showFilter');
+          tpColumn.toggleProperty('showFilter')
         }
-      });
+      })
     }
   }
-});
+})