+++ addon/components/table-pagination/table-footer.js
@@ -1,6 +1,6 @@
-import Component from '@ember/component';
-import layout from '../../templates/components/table-pagination/table-footer';
+import Component from '@ember/component'
+import layout from '../../templates/components/table-pagination/table-footer'
 
 export default Component.extend({
   layout
-});
+})