+++ addon/helpers/height-value.js
@@ -1,8 +1,8 @@
-import { helper } from '@ember/component/helper';
-import { htmlSafe } from '@ember/string';
+import { helper } from '@ember/component/helper'
+import { htmlSafe } from '@ember/string'
 
-export function heightValue([height]) {
-  return htmlSafe(`height: ${height}`);
+export function heightValue ([height]) {
+  return htmlSafe(`height: ${height}`)
 }
 
-export default helper(heightValue);
+export default helper(heightValue)