+++ app/components/bs-table-pagination/table-footer.js
@@ -1,1 +1,1 @@
-export {default} from 'ember-cli-table-pagination/components/bs-table-pagination/table-footer';
+export {default} from 'ember-cli-table-pagination/components/bs-table-pagination/table-footer'