+++ app/components/light-table-pagination.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-cli-table-pagination/components/light-table-pagination';
+export { default } from 'ember-cli-table-pagination/components/light-table-pagination'