+++ app/components/light-table-pagination/table-actions.js
@@ -1,2 +1,1 @@
-export {default} from 'ember-cli-table-pagination/components/light-table-pagination/table-actions';
-
+export {default} from 'ember-cli-table-pagination/components/light-table-pagination/table-actions'