+++ app/components/light-table/columns/base-with-filter.js
@@ -1,1 +1,1 @@
-export { default } from 'ember-cli-table-pagination/components/light-table/columns/base-with-filter';
+export { default } from 'ember-cli-table-pagination/components/light-table/columns/base-with-filter'